+++ src/psd_tools/user_api/shape.py
@@ -213,3 +213,124 @@ class VectorMask(object):
         """List of vertices of all subpaths."""
         return [p['anchor'] for p in self._setting.path
                 if p.get('selector') in self._KNOT_KEYS]
+
+
+class Origination(object):
+    """Vector origination.
+
+    Vector origination keeps live shape properties such as bounding box.
+    """
+    def __init__(self, data):
+        self._descriptor = data
+
+    @property
+    def origin_type(self):
+        """Type of the origination.
+
+        * 1: rectangle
+        * 2: rounded rectangle
+        * 4: line
+        * 5: ellipse
+
+        :rtype: int
+        """
+        return self._descriptor.get(b'keyOriginType')
+
+    @property
+    def resolution(self):
+        """Resolution.
+
+        :rtype: float
+        """
+        return self._descriptor.get(b'keyOriginResolution')
+
+    @property
+    def radii(self):
+        """Corner radii of rounded rectangles.
+        The order is top-left, top-right, bottom-left, bottom-right.
+
+        :rtype: tuple of float
+        """
+        return self._descriptor.get(b'keyOriginRRectRadii')
+
+    @property
+    def shape_bbox(self):
+        """Bounding box of the live shape.
+
+        :rtype: BBox
+        """
+        return self._descriptor.get(b'keyOriginShapeBBox')
+
+    @property
+    def line_end(self):
+        """Line end.
+
+        :rtype: Point
+        """
+        return self._descriptor.get(b'keyOriginLineEnd')
+
+    @property
+    def line_start(self):
+        """Line start.
+
+        :rtype: Point
+        """
+        return self._descriptor.get(b'keyOriginLineStart')
+
+    @property
+    def line_weight(self):
+        """Line weight
+
+        :rtype: float
+        """
+        return self._descriptor.get(b'keyOriginLineWeight')
+
+    @property
+    def arrow_start(self):
+        """Line arrow start.
+
+        :rtype: bool
+        """
+        return self._descriptor.get(b'keyOriginLineArrowSt')
+
+    @property
+    def arrow_end(self):
+        """Line arrow end.
+
+        :rtype: bool"""
+        return self._descriptor.get(b'keyOriginLineArrowEnd')
+
+    @property
+    def arrow_width(self):
+        """Line arrow width.
+
+        :rtype: float
+        """
+        return self._descriptor.get(b'keyOriginLineArrWdth')
+
+    @property
+    def arrow_length(self):
+        """Line arrow length.
+
+        :rtype: float
+        """
+        return self._descriptor.get(b'keyOriginLineArrLngth')
+
+    @property
+    def arrow_conc(self):
+        """
+
+        :rtype: int
+        """
+        return self._descriptor.get(b'keyOriginLineArrConc')
+
+    @property
+    def index(self):
+        """
+
+        :rtype: int
+        """
+        return self._descriptor.get(b'keyOriginIndex')
+
+    def __repr__(self):
+        return self._descriptor.__repr__()