+++ tests/test_info.py
@@ -17,22 +17,22 @@ FILES_WITH_NO_LAYERS = (
     ('32bit.psd',           True),
     ('32bit5x5.psd',        True),
     ('300dpi.psd',          True),
-    ('gradient fill.psd',   True),
+    ('gradient-fill.psd',   True),
     ('history.psd',         True),
     ('pen-text.psd',        True),
     ('transparentbg.psd',   True),
-    ('vector mask.psd',     True),
+    ('vector-mask.psd',     True),
     ('0layers.psb',         True),
     ('0layers_tblocks.psb', True),
     ('16bit5x5.psb',        True),
     ('32bit.psb',           True),
     ('32bit5x5.psb',        True),
     ('300dpi.psb',          True),
-    ('gradient fill.psb',   True),
+    ('gradient-fill.psb',   True),
     ('history.psb',         True),
     ('pen-text.psb',        True),
     ('transparentbg.psb',   True),
-    ('vector mask.psb',     True)
+    ('vector-mask.psb',     True)
 )
 
 
@@ -107,7 +107,7 @@ def test_api():
 
 
 def test_vector_mask():
-    psd = decode_psd('vector mask.psd')
+    psd = decode_psd('vector-mask.psd')
     layers = psd.layer_and_mask_data.layers.layer_records
     assert layers[1].tagged_blocks[1].key == TaggedBlock.VECTOR_MASK_SETTING1
     assert isinstance(layers[1].tagged_blocks[1].data, VectorMaskSetting)
@@ -129,10 +129,34 @@ def test_shape_paths():
     assert path.closed
 
 
-def test_vector_stroke_content_setting():
+@pytest.fixture(scope='module')
+def stroke_psd():
     psd = PSDImage(decode_psd('stroke.psd'))
-    assert psd.layers[1].kind == 'shape'
-    assert isinstance(psd.layers[1].stroke_content, PatternOverlay)
+    yield psd
+
+
+def test_vector_stroke_content_setting(stroke_psd):
+    assert stroke_psd.layers[1].kind == 'shape'
+    assert isinstance(stroke_psd.layers[1].stroke_content, PatternOverlay)
+
+
+def test_vector_origination(stroke_psd):
+    assert stroke_psd.layers[0].has_origination
+    origination = stroke_psd.layers[0].origination
+    assert origination.origin_type == 4
+    assert origination.resolution == 150.0
+    assert origination.shape_bbox == (187.0, 146.0, 220.0, 206.0)
+    assert origination.line_end.x == 220.0
+    assert origination.line_end.y == 146.0
+    assert origination.line_start.x == 187.0
+    assert origination.line_start.y == 206.0
+    assert origination.line_weight == 1.0
+    assert origination.arrow_start == False
+    assert origination.arrow_end == False
+    assert origination.arrow_width == 0.0
+    assert origination.arrow_length == 0.0
+    assert origination.arrow_conc == 0
+    assert origination.index == 0
 
 
 def test_print():