+++ tests/test_pixels.py
@@ -41,13 +41,13 @@ PIXEL_COLORS_32BIT = with_psb((
     ('32bit.psd',               (95, 15),     (0, 0, 0)),
     ('300dpi.psd',              (70, 30),     (0, 0, 0)),
     ('300dpi.psd',              (50, 60),     (214, 59, 59)),
-    ('gradient fill.psd',       (10, 15),     (235, 241, 250)), # background
-    ('gradient fill.psd',       (70, 50),     (0, 0, 0)), # black circle
-    ('gradient fill.psd',       (50, 50),     (205, 144, 110)), # filled ellipse
+    ('gradient-fill.psd',       (10, 15),     (235, 241, 250)), # background
+    ('gradient-fill.psd',       (70, 50),     (0, 0, 0)), # black circle
+    ('gradient-fill.psd',       (50, 50),     (205, 144, 110)), # filled ellipse
     ('pen-text.psd',            (50, 50),     (229, 93, 93)),
     ('pen-text.psd',            (170, 40),    (0, 0, 0)),
-    ('vector mask.psd',         (10, 15),     (255, 255, 255)),
-    ('vector mask.psd',         (50, 90),     (221, 227, 236)),
+    ('vector-mask.psd',         (10, 15),     (255, 255, 255)),
+    ('vector-mask.psd',         (50, 90),     (221, 227, 236)),
     ('transparentbg.psd',       (0, 0),       (255, 255, 255, 0)),
     ('transparentbg.psd',       (50, 50),     (0, 0, 0, 255)),
     ('32bit5x5.psd',            (0, 0),       (235, 241, 250)), # why not equal to 16bit5x5.psd?